add: make call-center the same in room and calls
