+++ tauri/src/components/ui/call-center.tsx
@@ -1,5 +1,5 @@
 import { formatDistanceToNow } from "date-fns";
-import { LuMicOff, LuVideo, LuVideoOff, LuScreenShare, LuScreenShareOff, LuWifiOff } from "react-icons/lu";
+import { LuMicOff, LuVideo, LuVideoOff, LuScreenShare, LuScreenShareOff } from "react-icons/lu";
 import { PiScribbleLoopBold, PiCursorBold } from "react-icons/pi";
 import useStore, { CallState, ParticipantRole } from "@/store/store";
 import { Separator } from "@/components/ui/separator";
@@ -10,12 +10,11 @@ import {
   DropdownMenuTrigger,
   DropdownMenuCheckboxItem,
 } from "@/components/ui/dropdown-menu";
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { Select, SelectContent, SelectItem, SelectTrigger } from "./select";
 import { SelectPortal } from "@radix-ui/react-select";
 import { Button } from "./button";
 import { tauriUtils } from "@/windows/window-utils";
-import { HoppAvatar } from "./hopp-avatar";
 import { HiOutlineCursorClick, HiOutlineEye } from "react-icons/hi";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip";
 import { CustomIcons } from "@/components/ui/icons";
@@ -28,6 +27,7 @@ import { useEndCall } from "@/lib/hooks";
 import { typedInvoke } from "@/core_payloads";
 import { useQuery } from "@tanstack/react-query";
 import { listen } from "@tauri-apps/api/event";
+import { sounds } from "@/constants/sounds";
 
 const Colors = {
   deactivatedIcon: "text-slate-600",
@@ -39,7 +39,6 @@ const Colors = {
 
 export function CallCenter() {
   const { callTokens, teammates } = useStore();
-  const callParticipant = teammates?.find((user) => user.id === callTokens?.participant);
   const [controllerCursorState, setControllerCursorState] = useState(true);
   const [accessibilityPermission, setAccessibilityPermission] = useState(true);
 
@@ -52,13 +51,6 @@ export function CallCenter() {
   const handleEndCall = useEndCall();
   const handleEndCallRef = useRef(handleEndCall);
   handleEndCallRef.current = handleEndCall;
-
-  // Determine remote participant state from core events
-  const remoteParticipantState = callTokens?.participants?.find(
-    (p) => callParticipant && p.identity.includes(callParticipant.id),
-  );
-  const isRemoteDisconnected = remoteParticipantState ? !remoteParticipantState.connected : false;
-  const isRemoteMuted = remoteParticipantState?.muted ?? false;
 
   const fetchAccessibilityPermission = async () => {
     const permission = await tauriUtils.getControlPermission();
@@ -135,126 +127,190 @@ export function CallCenter() {
           </div>
         )}
       </div>
-      <div
-        className={clsx("gap-2 px-3 flex-nowrap grid mb-4 min-w-full", {
-          "grid-cols-6": callTokens?.isRoomCall,
-          "grid-cols-12": !callTokens?.isRoomCall,
-        })}
-      >
-        {!callTokens?.isRoomCall && (
-          <div className="flex flex-col items-start mb-4 col-span-3 relative">
-            <div className="relative mt-1">
-              {callParticipant && (
-                <>
-                  <HoppAvatar
-                    src={callParticipant?.avatar_url || undefined}
-                    firstName={callParticipant?.first_name}
-                    lastName={callParticipant?.last_name}
-                    isMuted={isRemoteMuted}
-                  />
-                  {isRemoteDisconnected && (
-                    <div className="absolute -top-1 -right-1 bg-red-500 rounded-full p-1 shadow-md border-2 border-white">
-                      <LuWifiOff className="size-3 text-white" />
-                    </div>
-                  )}
-                </>
-              )}
-            </div>
-            <div className="flex flex-col items-start mt-2 w-full">
-              <span className="text-sm text-left w-full">{callParticipant?.first_name}</span>
-              <span className="text-sm text-left text-slate-500 w-full truncate">{callParticipant?.last_name}</span>
-            </div>
-          </div>
-        )}
-        <div className="flex flex-col gap-2 items-center col-span-9">
-          <div className="flex flex-row gap-1 w-full">
-            <MediaDevicesSettings
-              micShortcut={userSettings?.shortcut_toggle_mic}
-              cameraShortcut={userSettings?.shortcut_toggle_camera}
-              screenShareShortcut={userSettings?.shortcut_toggle_screenshare}
-            />
-          </div>
-          <div className="flex flex-col gap-2 w-full">
-            {callTokens?.role === ParticipantRole.CONTROLLER && (
-              <Button
-                className="w-full border-gray-500 text-gray-600 flex flex-row gap-2"
-                variant="gradient-white"
-                onClick={() => {
-                  typedInvoke("open_screenshare_viewer");
-                }}
-              >
-                <HiOutlineEye className="size-4" />
-                Open shared window
-              </Button>
-            )}
-            <div className="w-full flex flex-row gap-2">
-              {callTokens?.role === ParticipantRole.SHARER && (
-                <TooltipProvider>
-                  <Tooltip delayDuration={100}>
-                    <TooltipTrigger>
-                      <ToggleIconButton
-                        onClick={() => {
-                          let controllerCursorTmp = !controllerCursorState;
-                          tauriUtils.setControllerCursor(controllerCursorTmp);
-                          setControllerCursorState(controllerCursorTmp);
-                        }}
-                        state={
-                          controllerCursorState ? "active"
-                          : !accessibilityPermission ?
-                            "deactivated"
-                          : "neutral"
-                        }
-                        size="unsized"
-                        className="size-9"
-                        disabled={!accessibilityPermission}
-                      >
-                        {controllerCursorState && (
-                          <HiOutlineCursorClick
-                            className={clsx("size-4", {
-                              "text-red-800": !controllerCursorState,
-                            })}
-                          />
-                        )}
-                        {!controllerCursorState && (
-                          <div className="relative">
-                            <HiOutlineCursorClick className="size-4 text-red-800" />
-                            <span className="absolute bottom-[-8px] right-[-7px] text-[10px]">🔒</span>
-                          </div>
-                        )}
-                      </ToggleIconButton>
-                    </TooltipTrigger>
-                    <TooltipContent side="bottom">
-                      {controllerCursorState ? "Disable" : "Enable"} remote control
-                    </TooltipContent>
-                  </Tooltip>
-                </TooltipProvider>
-              )}
-              {callTokens?.role === ParticipantRole.SHARER && <DrawingEnableButton />}
+      <div className="gap-2 px-3 flex flex-col items-center w-full mb-2">
+        <div className="flex flex-row gap-1 w-full">
+          <MediaDevicesSettings
+            micShortcut={userSettings?.shortcut_toggle_mic}
+            cameraShortcut={userSettings?.shortcut_toggle_camera}
+            screenShareShortcut={userSettings?.shortcut_toggle_screenshare}
+          />
+        </div>
+        <div className="flex flex-col gap-2 w-full">
+          {callTokens?.role === ParticipantRole.CONTROLLER && (
+            <Button
+              className="w-full border-gray-500 text-gray-600 flex flex-row gap-2"
+              variant="gradient-white"
+              onClick={() => {
+                typedInvoke("open_screenshare_viewer");
+              }}
+            >
+              <HiOutlineEye className="size-4" />
+              Open shared window
+            </Button>
+          )}
+          <div className="w-full flex flex-row gap-2">
+            {callTokens?.role === ParticipantRole.SHARER && (
               <TooltipProvider>
                 <Tooltip delayDuration={100}>
-                  <TooltipTrigger asChild>
-                    <Button
-                      className="w-full border-red-500 text-red-600 flex flex-row gap-2"
-                      variant="gradient-white"
-                      onClick={handleEndCall}
+                  <TooltipTrigger>
+                    <ToggleIconButton
+                      onClick={() => {
+                        let controllerCursorTmp = !controllerCursorState;
+                        tauriUtils.setControllerCursor(controllerCursorTmp);
+                        setControllerCursorState(controllerCursorTmp);
+                      }}
+                      state={
+                        controllerCursorState ? "active"
+                        : !accessibilityPermission ?
+                          "deactivated"
+                        : "neutral"
+                      }
+                      size="unsized"
+                      className="size-9"
+                      disabled={!accessibilityPermission}
                     >
-                      <HiOutlinePhoneXMark className="size-4" />
-                      End call
-                    </Button>
+                      {controllerCursorState && (
+                        <HiOutlineCursorClick
+                          className={clsx("size-4", {
+                            "text-red-800": !controllerCursorState,
+                          })}
+                        />
+                      )}
+                      {!controllerCursorState && (
+                        <div className="relative">
+                          <HiOutlineCursorClick className="size-4 text-red-800" />
+                          <span className="absolute bottom-[-8px] right-[-7px] text-[10px]">🔒</span>
+                        </div>
+                      )}
+                    </ToggleIconButton>
                   </TooltipTrigger>
-                  {userSettings?.shortcut_end_call && (
-                    <TooltipContent side="bottom" variant="light">
-                      <span className="font-mono text-xs">{userSettings.shortcut_end_call}</span>
-                    </TooltipContent>
-                  )}
+                  <TooltipContent side="bottom">
+                    {controllerCursorState ? "Disable" : "Enable"} remote control
+                  </TooltipContent>
                 </Tooltip>
               </TooltipProvider>
-            </div>
+            )}
+            {callTokens?.role === ParticipantRole.SHARER && <DrawingEnableButton />}
+            <TooltipProvider>
+              <Tooltip delayDuration={100}>
+                <TooltipTrigger asChild>
+                  <Button
+                    className="w-full border-red-500 text-red-600 flex flex-row gap-2"
+                    variant="gradient-white"
+                    onClick={handleEndCall}
+                  >
+                    <HiOutlinePhoneXMark className="size-4" />
+                    End call
+                  </Button>
+                </TooltipTrigger>
+                {userSettings?.shortcut_end_call && (
+                  <TooltipContent side="bottom" variant="light">
+                    <span className="font-mono text-xs">{userSettings.shortcut_end_call}</span>
+                  </TooltipContent>
+                )}
+              </Tooltip>
+            </TooltipProvider>
           </div>
         </div>
       </div>
+      <CallParticipants />
       {/* Horizontal line */}
-      <Separator className="w-full" />
+      <Separator className="w-full mt-4" />
+    </div>
+  );
+}
+
+function CallParticipants() {
+  const { callTokens, teammates, user } = useStore();
+  const coreParticipants = callTokens?.participants ?? [];
+  const prevCountRef = useRef(coreParticipants.length);
+
+  // Play sound when a new participant connects (count increases)
+  useEffect(() => {
+    if (coreParticipants.length > prevCountRef.current && prevCountRef.current !== 0) {
+      sounds.callAccepted.play();
+    }
+    prevCountRef.current = coreParticipants.length;
+  }, [coreParticipants.length]);
+
+  const participantList = useMemo(() => {
+    const extractUserId = (identity: string): string => {
+      const parts = identity.split(":");
+      return parts.length >= 4 ? (parts[2] ?? identity) : identity;
+    };
+
+    const findUser = (participantId: string) => {
+      if (user && user.id === participantId) return user;
+      return teammates?.find((t) => t.id === participantId) ?? null;
+    };
+
+    const remoteEntries = coreParticipants
+      .filter((p) => p.connected)
+      .filter((p) => {
+        if (p.identity === "local") return false;
+        const pid = extractUserId(p.identity);
+        return pid !== user?.id;
+      })
+      .map((p) => {
+        const participantId = extractUserId(p.identity);
+        return {
+          id: p.identity,
+          participantId,
+          user: findUser(participantId),
+          isMicrophoneEnabled: !p.muted,
+        };
+      });
+
+    return remoteEntries;
+  }, [coreParticipants, teammates, user]);
+
+  if (participantList.length === 0) return null;
+
+  return (
+    <div className="flex flex-col w-full mt-3 px-3">
+      <span className="text-xs font-medium text-slate-600 mb-2">Participants</span>
+      <div className="flex flex-col gap-3">
+        {participantList.map((participant) => (
+          <div key={participant.id} className="flex items-center gap-3">
+            {participant.user ?
+              <>
+                <div className="size-8 rounded-md bg-emerald-200 flex items-center justify-center shrink-0 overflow-hidden ring-1 ring-white">
+                  {participant.user.avatar_url ?
+                    <img
+                      src={participant.user.avatar_url}
+                      alt={`${participant.user.first_name} ${participant.user.last_name}`}
+                      className="size-full object-cover"
+                    />
+                  : <span className="text-[10px] font-medium text-emerald-700">
+                      {participant.user.first_name[0]}
+                      {participant.user.last_name[0]}
+                    </span>
+                  }
+                </div>
+                <div className="flex flex-col">
+                  <span className="text-sm font-medium">
+                    {participant.user.first_name} {participant.user.last_name}
+                  </span>
+                  {!participant.isMicrophoneEnabled && (
+                    <span className="flex items-center gap-1 text-xs font-medium text-orange-500">
+                      <LuMicOff className="size-3" />
+                      <span className="mt-0.5">Muted</span>
+                    </span>
+                  )}
+                </div>
+              </>
+            : <>
+                <div className="size-8 rounded-md bg-slate-200 flex items-center justify-center shrink-0">
+                  <span className="text-xs font-medium text-slate-600">?</span>
+                </div>
+                <div className="flex flex-col">
+                  <span className="text-sm font-medium text-slate-600">Unknown user</span>
+                </div>
+              </>
+            }
+          </div>
+        ))}
+      </div>
     </div>
   );
 }
